Rename files to match naming convention and re-organise util files into folders
