+++ src/components/SideBarMenu.tsx
@@ -1,12 +1,11 @@
-// import CreateNewRec from './createnewrec';
-import "../css/sidebar.css";
-import UsersFetch from "../utils/UsersFetch";
-import TypeFetch from "../utils/TypeFetch";
+import "../css/SideBarMenu.css";
+import UsersFetch from "../utils/Helper-Functions/UsersFetch";
+import TypeFetch from "../utils/Helper-Functions/TypeFetch";
 import NewModal from "./NewModal";
 import { useState, useEffect } from "react";
-import separateCapitalise from "../utils/separateCapitalise";
+import separateCapitalise from "../utils/Helper-Functions/separateCapitalise";
 import { Link } from "react-router-dom";
-import fetchTags from "../utils/TagsFetch";
+import fetchTags from "../utils/Helper-Functions/TagsFetch";
 
 interface SidebarProps {
   currentUser: number;
@@ -105,7 +104,6 @@ export default function SideBarMenu(props: SidebarProps): JSX.Element {
         </form>
         <br id="inner" />
         <br id="inner" />
-        {/* <form> */}
         <h3>Tags</h3>
         {tagCloud}
         {props.selectedTags.length !== 0 && (
@@ -119,8 +117,6 @@ export default function SideBarMenu(props: SidebarProps): JSX.Element {
             <button>OK</button>
           </Link>
         )}
-
-        {/* </form> */}
         <br id="inner" />
         <br id="inner" />
         {props.currentUser === 0 && (
