--- src/components/tagscloud.tsx
@@ -1,3 +1,0 @@
-export default function TagsCloud(): JSX.Element {
-  return <></>;
-}
