+++ src/utils/Props/CommentProps.tsx
@@ -6,5 +6,3 @@ export default interface CommentProps {
   comment: string;
   name: string;
 }
-
-// id":2,"user_id":2,"rec_id":1,"submit_time":"2022-01-04T16:05:34.181Z","comment":"I hated this, never reading this again","name":"Hanna"}
