+++ src/utils/Props/StudyListProps.tsx
@@ -1,0 +1,4 @@
+export interface PostStudyListProps {
+  user_id: number;
+  rec_id: number;
+}
