--- src/utils/StudyListProps.tsx
@@ -1,14 +1,0 @@
-export interface PostStudyListProps {
-  user_id: number;
-  rec_id: number;
-}
-
-// export interface GetStudyListProps {
-//     id: number,
-//     title: string,
-//     author: string,
-//     type: string,
-//     summary: string,
-//     link: string,
-//     submit_time: string,
-// }
