+++ src/utils/Tests/separateCapitalise.test.ts
@@ -1,4 +1,4 @@
-import separateCapitalise from "./separateCapitalise";
+import separateCapitalise from "../Helper-Functions/separateCapitalise";
 
 test("separateCapitalise returns the input string with the first letter capitalised", () => {
   expect(separateCapitalise("hello world")).toStrictEqual("Hello World");
